Add SolidusSupport.payment_method_parent_class
This helps to get the parent class for payment methods.

For Solidus < 2.3 it is Spree::PaymentMethod for non credit card
payment methods or Spree::Gateway for credit card payment methods.

For Solidus >= 2.3 it is Spree::PaymentMethod for non credit card
payment methods or Spree::PaymentMethod::CreditCard for credit card
payment methods.

Usage:

    My::NonCreditCardPaymentMethod < SolidusSupport.payment_method_parent_class

    My::CreditCardPaymentMethod <
    SolidusSupport.payment_method_parent_class(credit_card: true)