Add a method to check if combined name is being used
This helper method checks if a store is using the old or new
way of setting address' name.

It will act as a migration support to create conditional code
to make extensions work compatible with both versions.

We can remove this method when the last 2.x version will reach EOL.