Allow detecting Solidus components through a Rails setting
This allows us to define solidus components at the application level, in
contrast to the typical engine definition.

Components paths are loaded on an initializer, which runs before the
application code that could define an `Engine` class. Adding a setting
on the `application.rb`, which runs before the initializers, helps us
working around it.

For now, we keep the old detection mechanism, as the other extensions
will need to adapt to it first.