Feat(Performance): Use Zeitwerk for loading decorators
This is inspired by #60.

We can leverage Zeitwerk's `on_load` hook and it's capacity of knowing
which constant a file should define in order to load decorators,
including when reloading.

This should greatly speed up reloading, as only those decorators that
are needed for the current request are loaded.

However, there is a few restrictions that come with this:

1. All decorators MUST use a Zeitwerk-compatible naming scheme
2. All decorators MUST use Module.prepend, where Module is the fully
   qualified class name being modified.

Co-Authored-By: thomas@vondeyen.com