Support Flickwerk-style patches
This enables using patches (i.e. "decorators"/"overrides") in Solidus
engines, but using the style described in the Flickwerk README[^1]

In short, this allows autoloading patches whenever the file being
patches is loaded by Zeitwerk.

The new directory to put your patches into would be `app/patches`, and
the files inside should be named `my_extension/spree_user_patch.rb` and
define a constant `MyExtension::SpreeUserPatch`. The convention is that
the module prepends itself to the class being patched, in this example
`Spree::User`.

[^1] https://github.com/friendlycart/flickwerk/blob/main/README.md