Add missing |app| argument, remove sorting specs in initializers
Without the missing |app| argument after the second initializer, loading
fails with "app" being the engine's route set, not the actual app.

The `before` and `after` specifications are not necessary, as they are
not dependent on load order. Either they just add some paths to
Flickwerk if they exist, which is entirely fine, or they add things to
the `app.reloader.to_prepare` callback, which will be executed at the
right time.